moves archive logic in gzip file
